changed the integration and with complex behaviour data objects to adapters
