+++ .rubocop.yml
@@ -45,9 +45,9 @@ Metrics/ClassLength:
   Exclude:
     - 'app/models/project.rb'
     - 'app/repositories/project_results_repository.rb'
-    - 'app/services/pipefy/pipefy_card_response_reader.rb'
-    - 'app/data_objects/report_data.rb'
-    - 'app/data_objects/status_report_data.rb'
+    - 'app/adapters/pipefy/pipefy_card_adapter.rb'
+    - 'app/adapters/highchart/operational_charts_adapter.rb'
+    - 'app/adapters/highchart/status_report_charts_adapter.rb'
     - 'app/models/demand.rb'
     - 'app/controllers/projects_controller.rb'
