+++ app/controllers/charts_controller.rb
@@ -5,17 +5,17 @@ class ChartsController < AuthenticatedController
   before_action :assign_projects
 
   def build_operational_charts
-    @report_data = ReportData.new(@projects)
+    @report_data = Highchart::OperationalChartsAdapter.new(@projects)
     respond_to { |format| format.js { render file: 'teams/operational_charts.js.erb' } }
   end
 
   def build_strategic_charts
-    @strategic_report_data = StrategicReportData.new(@company, @projects, @available_hours_in_month)
+    @strategic_report_data = Highchart::StrategicChartsAdapter.new(@company, @projects, @available_hours_in_month)
     respond_to { |format| format.js { render file: 'teams/strategic_charts.js.erb' } }
   end
 
   def build_status_report_charts
-    @status_report_data = StatusReportData.new(@projects)
+    @status_report_data = Highchart::StatusReportChartsAdapter.new(@projects)
     respond_to { |format| format.js { render file: 'teams/status_report_charts.js.erb' } }
   end
