+++ app/controllers/products_controller.rb
@@ -10,9 +10,9 @@ class ProductsController < AuthenticatedController
 
   def show
     @product_projects = @product.projects.order(end_date: :desc)
-    @projects_summary = ProjectsSummaryObject.new(@product.projects)
-    @report_data = ReportData.new(@product_projects) if @product_projects.present?
-    @status_report_data = StatusReportData.new(@product_projects) if @product_projects.present?
+    @projects_summary = ProjectsSummaryData.new(@product.projects)
+    @report_data = Highchart::OperationalChartsAdapter.new(@product_projects) if @product_projects.present?
+    @status_report_data = Highchart::StatusReportChartsAdapter.new(@product_projects) if @product_projects.present?
   end
 
   def new
@@ -46,7 +46,7 @@ class ProductsController < AuthenticatedController
 
   def search_for_projects
     @projects = ProjectsRepository.instance.add_query_to_projects_in_status(@product.projects, params[:status_filter])
-    @projects_summary = ProjectsSummaryObject.new(@projects)
+    @projects_summary = ProjectsSummaryData.new(@projects)
     respond_to { |format| format.js { render file: 'projects/projects_search.js.erb' } }
   end
