+++ spec/adapters/highchart/burnup_charts_adapter_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-RSpec.describe BurnupData, type: :data_object do
+RSpec.describe Highchart::BurnupChartsAdapter, type: :data_object do
   context 'having projects' do
     describe '.initialize' do
-      subject(:burnup_data) { BurnupData.new([12, 13, 14, 15], [2, 4, 5, 6], [1, 5, 5, 10]) }
+      subject(:burnup_data) { Highchart::BurnupChartsAdapter.new([12, 13, 14, 15], [2, 4, 5, 6], [1, 5, 5, 10]) }
 
       it 'do the math and provides the correct information' do
         expect(burnup_data.burnup_x_axis_period).to eq [12, 13, 14, 15]
@@ -16,7 +16,7 @@ RSpec.describe BurnupData, type: :data_object do
 
   context 'having no data' do
     describe '.initialize' do
-      subject(:burnup_data) { BurnupData.new([], [], []) }
+      subject(:burnup_data) { Highchart::BurnupChartsAdapter.new([], [], []) }
 
       it 'returns empty arrays' do
         expect(burnup_data.burnup_x_axis_period).to eq []
