+++ spec/adapters/highchart/strategic_charts_adapter_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe StrategicReportData, type: :service do
+RSpec.describe Highchart::StrategicChartsAdapter, type: :service do
   before { travel_to Time.zone.local(2018, 2, 20, 10, 0, 0) }
   after { travel_back }
 
@@ -36,7 +36,7 @@ RSpec.describe StrategicReportData, type: :service do
       let!(:eighth_project_result) { Fabricate :project_result, project: eighth_project, team: team, result_date: 2.months.from_now, qty_hours_upstream: 10, qty_hours_downstream: 40, available_hours: 234, throughput_upstream: 5, throughput_downstream: 18, flow_pressure: 1 }
 
       it 'mounts the data structure to the active project counts in months' do
-        strategic_data = StrategicReportData.new(company, company.projects, company.total_available_hours)
+        strategic_data = Highchart::StrategicChartsAdapter.new(company, company.projects, company.total_available_hours)
         expect(strategic_data.array_of_months).to eq [[3.months.ago.to_date.month, 3.months.ago.to_date.year], [2.months.ago.to_date.month, 2.months.ago.to_date.year], [1.month.ago.to_date.month, 1.month.ago.to_date.year], [Time.zone.today.month, Time.zone.today.year], [1.month.from_now.to_date.month, 1.month.from_now.to_date.year], [2.months.from_now.to_date.month, 2.months.from_now.to_date.year], [3.months.from_now.to_date.month, 3.months.from_now.to_date.year]]
         expect(strategic_data.active_projects_count_data).to eq [2, 2, 0, 0, 2, 4, 2]
         expect(strategic_data.sold_hours_in_month).to eq [1451.6129032258063, 1451.6129032258063, 0, 0, 2062.5, 7094.758064516129, 5032.258064516129]
@@ -50,7 +50,7 @@ RSpec.describe StrategicReportData, type: :service do
 
     context 'having no projects' do
       it 'returns an empty array' do
-        strategic_data = StrategicReportData.new(company, company.projects, company.total_available_hours)
+        strategic_data = Highchart::StrategicChartsAdapter.new(company, company.projects, company.total_available_hours)
         expect(strategic_data.array_of_months).to eq []
         expect(strategic_data.active_projects_count_data).to eq []
       end
