+++ spec/controllers/customers_controller_spec.rb
@@ -188,8 +188,8 @@ RSpec.describe CustomersController, type: :controller do
             expect(response).to render_template :show
             expect(assigns(:company)).to eq company
             expect(assigns(:customer)).to eq customer
-            expect(assigns(:report_data)).to be_a ReportData
-            expect(assigns(:status_report_data)).to be_a StatusReportData
+            expect(assigns(:report_data)).to be_a Highchart::OperationalChartsAdapter
+            expect(assigns(:status_report_data)).to be_a Highchart::StatusReportChartsAdapter
             expect(assigns(:customer_projects)).to eq [second_project, first_project]
           end
         end
@@ -199,8 +199,8 @@ RSpec.describe CustomersController, type: :controller do
             expect(response).to render_template :show
             expect(assigns(:company)).to eq company
             expect(assigns(:customer)).to eq customer
-            expect(assigns(:report_data)).to be_a ReportData
-            expect(assigns(:status_report_data)).to be_a StatusReportData
+            expect(assigns(:report_data)).to be_a Highchart::OperationalChartsAdapter
+            expect(assigns(:status_report_data)).to be_a Highchart::StatusReportChartsAdapter
             expect(assigns(:customer_projects)).to eq []
           end
         end
