+++ spec/data_objects/projects_summary_data_spec.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-describe ProjectsSummaryObject, type: :data_object do
+describe ProjectsSummaryData, type: :data_object do
   describe '#percentage_remaining_money' do
     context 'when the total hours is different of zero' do
       let!(:project) { Fabricate :project }
       let!(:project_result) { Fabricate :project_result, project: project }
       let!(:other_project_result) { Fabricate :project_result, project: project }
 
-      subject(:projects_summary) { ProjectsSummaryObject.new(Project.all) }
+      subject(:projects_summary) { ProjectsSummaryData.new(Project.all) }
 
       it { expect(projects_summary.percentage_remaining_money).to eq((projects_summary.total_remaining_money.to_f / projects_summary.total_value.to_f) * 100) }
     end
@@ -19,7 +19,7 @@ describe ProjectsSummaryObject, type: :data_object do
       let!(:second_project) { Fabricate :project, value: 0 }
       let!(:second_project_result) { Fabricate :project_result, project: second_project }
 
-      subject(:projects_summary) { ProjectsSummaryObject.new(Project.all) }
+      subject(:projects_summary) { ProjectsSummaryData.new(Project.all) }
 
       it { expect(projects_summary.percentage_remaining_money).to eq 0 }
     end
@@ -34,7 +34,7 @@ describe ProjectsSummaryObject, type: :data_object do
       let!(:second_project) { Fabricate :project }
       let!(:second_project_result) { Fabricate :project_result, project: second_project }
 
-      subject(:projects_summary) { ProjectsSummaryObject.new(Project.all) }
+      subject(:projects_summary) { ProjectsSummaryData.new(Project.all) }
 
       it { expect(projects_summary.percentage_remaining_days).to eq((projects_summary.total_remaining_days.to_f / projects_summary.total_days.to_f) * 100) }
     end
@@ -43,7 +43,7 @@ describe ProjectsSummaryObject, type: :data_object do
       let!(:project_result) { Fabricate :project_result, project: project }
       let!(:other_project_result) { Fabricate :project_result, project: project }
 
-      subject(:projects_summary) { ProjectsSummaryObject.new(Project.all) }
+      subject(:projects_summary) { ProjectsSummaryData.new(Project.all) }
 
       it { expect(projects_summary.percentage_remaining_days).to eq 100 }
     end
@@ -58,7 +58,7 @@ describe ProjectsSummaryObject, type: :data_object do
       let!(:second_project) { Fabricate :project }
       let!(:second_project_result) { Fabricate :project_result, project: second_project }
 
-      subject(:projects_summary) { ProjectsSummaryObject.new(Project.all) }
+      subject(:projects_summary) { ProjectsSummaryData.new(Project.all) }
 
       it { expect(projects_summary.total_gap).to eq(projects_summary.total_last_week_scope - projects_summary.total_delivered_scope) }
     end
